[IMP] core: promote the json2 http.Dispatcher to core
Move the http.Dispatcher that was created for the new /json/2 endpoint
to core under the name "json2".

The Dispatcher is now mature enough and deserve to be promoted to core,
so other modules can use it as well.
